Cleaned up Warehousing code. Automatically determine umbrella/single
The umbrella/single determination system basically checks whether or not a given course title is represented >=50% of the time. If not, it assumes that the course title isn't standard i.e. has a bunch of variations (like in Special Topics courses). It then verifies that there's a common root label in the titles (e.g. "Special Topics"). If both of those things are true, it determines it as an Umbrella course.